requirements.txt
fastapi>=0.95.0
uvicorn>=0.21.1
pydantic>=1.10.7
python-jose>=3.3.0
passlib>=1.7.4
python-multipart>=0.0.6
sqlalchemy>=2.0.9
python-dotenv>=1.0.0
requests>=2.28.2
comet-ml>=3.33.0
python-dateutil>=2.8.2
pandas>=2.0.0
numpy>=1.24.2
pytesseract>=0.3.10
pdfplumber>=0.9.0
beautifulsoup4>=4.12.0
aiofiles>=23.1.0
pytest>=7.3.1
httpx>=0.24.0
